Include default voice samples

+++ src/chatterbox/models/s3gen/decoder.py
@@ -259,12 +259,12 @@ class ConditionalDecoder(nn.Module):
             _type_: _description_
         """
         t = t.to(dtype=self.dtype)
-        t = self.time_embeddings(t).to(t.dtype)
+        t = self.time_embeddings(t).to(dtype=self.dtype)
         t = self.time_mlp(t)
 
         if self.meanflow:
             r = r.to(dtype=self.dtype)
-            r = self.time_embeddings(r).to(t.dtype)
+            r = self.time_embeddings(r).to(dtype=self.dtype)
             r = self.time_mlp(r)
             concat_embed = torch.cat([t, r], dim=1)
             t = self.time_embed_mixer(concat_embed)
